agentic3d/pyproject.toml
[tool.poetry]
name = "agentic3d"
version = "0.1.0"
description = "AI Agent that develops OpenSCAD code and rendering based on initial object description, and iterative feedback improvements."
authors = ["Moritz Rietschel <rietschel@berkeley.edu>","Reet Mishra <reet.mishra@berkeley.edu>", "Sammie Shenon <sshenon@berkeley.edu>", "Frederik Stihler <frederik.stihler@berkeley.edu>"]
readme = "README.md"
packages = [{include = "agentic3d", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
autogen-agentchat = ">=0.2,<1.0"
pyautogen = "^0.5.3"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"

[virtualenvs]
create = true
in-project = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
